Use PouchDB Browser
PouchDB is able to run both in NodeJS as well as in a Browser. In a browser, IndexedDB will be used, for NodeJS, however, Leveldown is the backend, which installs a binary.

There are officially supported versions of PouchDB that don't come with this dependency (that fails to install on some modern versions of node due to prebuilds not being available). We should use these stripped versions to have less dependencies in our project.

See https://pouchdb.com/custom.html for more information.