Add a `Back` button to the final view of the attendance-addition form
Fixes #331

After being done with adding daily attendances, there's a view which
only contains a `Finish` button, but no `Back` button which is needed if
the user entered wrong data about the last child's attendance that needs
correction.

Also boldified the `Finish` button to make it visually even clearer which
button should be used to finish the action.