refactor: Remove that doesn’t run meaningful assertions
The removed test should check that dashboard includes widgets with
attendance summary. When you run the tests, you can see that no
attendance summary is visible because of the default fixtures don’t
include the necessary records. The tests never catch this because of
their dynamic nature. (`daysCount` is 0)

There are other issues with the tests (locators) and we don’t have a
good way to generate the necessary fixtures yet (see #2995). So instead
of replacing the test I opted to remove it for now.